Store agent files at workspace level
