+++ packages/server/src/sdk/index.ts
@@ -22,7 +22,7 @@ import * as views from "./workspace/views"
 import * as workspaceApps from "./workspace/workspaceApps"
 import { default as workspaces } from "./workspace/workspaces"
 import * as aiConfigs from "./aiConfig"
-import * as ragConfigs from "./ragConfig"
+import * as ragConfigs from "./workspace/ai/rag/configCrud"
 import * as vectorDbs from "./workspace/ai/vectorDb"
 
 const sdk = {
