+++ packages/server/src/sdk/workspace/ai/index.ts
@@ -1,4 +1,4 @@
 export * as agents from "./agents"
 export * as chatApps from "./chatApps"
 export * as helpers from "./helpers"
-export * as rag from "./rag"
+export * as rag from "./rag/files"
